requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --extra=dev setup.py
#
altair==4.1.0
    # via great-expectations
appnope==0.1.2
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.1.0
    # via notebook
attrs==21.2.0
    # via
    #   cattrs
    #   jsii
    #   jsonschema
    #   openlineage-airflow
    #   openlineage-integration-common
    #   openlineage-python
aws-cdk-lib==2.0.0rc29
    # via AwsCdkDatalake (setup.py)
awscli==1.22.5
    # via AwsCdkDatalake (setup.py)
backcall==0.2.0
    # via ipython
backports.zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
black==21.10b0
    # via AwsCdkDatalake (setup.py)
bleach==4.1.0
    # via nbconvert
boto3==1.20.5
    # via AwsCdkDatalake (setup.py)
botocore==1.23.5
    # via
    #   awscli
    #   boto3
    #   s3transfer
cachetools==4.2.4
    # via google-auth
cattrs==1.8.0
    # via jsii
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via argon2-cffi
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via
    #   black
    #   great-expectations
colorama==0.4.3
    # via awscli
constructs==10.0.9
    # via
    #   AwsCdkDatalake (setup.py)
    #   aws-cdk-lib
debugpy==1.5.1
    # via ipykernel
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via nbconvert
docutils==0.15.2
    # via awscli
entrypoints==0.3
    # via
    #   altair
    #   jupyter-client
    #   nbconvert
google-api-core[grpc]==2.2.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-bigquery==2.30.1
    # via AwsCdkDatalake (setup.py)
google-cloud-core==2.2.1
    # via google-cloud-bigquery
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.1.0
    # via google-cloud-bigquery
googleapis-common-protos==1.53.0
    # via
    #   google-api-core
    #   grpcio-status
great-expectations==0.13.42
    # via AwsCdkDatalake (setup.py)
grpcio==1.41.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.41.1
    # via google-api-core
idna==3.3
    # via requests
importlib-metadata==4.8.2
    # via great-expectations
importlib-resources==5.4.0
    # via jsonschema
ipykernel==6.5.0
    # via
    #   ipywidgets
    #   notebook
ipython==7.29.0
    # via
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   nbformat
    #   notebook
ipywidgets==7.6.5
    # via great-expectations
jedi==0.18.0
    # via ipython
jinja2==3.0.3
    # via
    #   altair
    #   great-expectations
    #   nbconvert
    #   notebook
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsii==1.43.0
    # via
    #   aws-cdk-lib
    #   constructs
jsonpatch==1.32
    # via great-expectations
jsonpointer==2.2
    # via jsonpatch
jsonschema==4.2.1
    # via
    #   altair
    #   great-expectations
    #   nbformat
jupyter-client==7.0.6
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.9.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.2
    # via ipywidgets
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via
    #   great-expectations
    #   nbconvert
mypy-extensions==0.4.3
    # via black
nbclient==0.5.8
    # via nbconvert
nbconvert==6.3.0
    # via notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via
    #   jupyter-client
    #   nbclient
notebook==6.4.5
    # via widgetsnbextension
numpy==1.21.4
    # via
    #   altair
    #   great-expectations
    #   pandas
    #   scipy
openlineage-airflow==0.3.1
    # via AwsCdkDatalake (setup.py)
openlineage-integration-common==0.3.1
    # via openlineage-airflow
openlineage-python==0.3.1
    # via
    #   AwsCdkDatalake (setup.py)
    #   openlineage-airflow
    #   openlineage-integration-common
packaging==21.2
    # via
    #   bleach
    #   google-cloud-bigquery
pandas==1.3.4
    # via
    #   altair
    #   great-expectations
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.2
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.4.0
    # via black
prometheus-client==0.12.0
    # via notebook
prompt-toolkit==3.0.22
    # via ipython
proto-plus==1.19.8
    # via google-cloud-bigquery
protobuf==3.19.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   constructs
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pygments==2.10.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pyparsing==2.4.7
    # via
    #   great-expectations
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   botocore
    #   google-cloud-bigquery
    #   great-expectations
    #   jsii
    #   jupyter-client
    #   pandas
pytz==2021.3
    # via
    #   great-expectations
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==5.4.1
    # via awscli
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
regex==2021.11.10
    # via black
requests==2.26.0
    # via
    #   AwsCdkDatalake (setup.py)
    #   google-api-core
    #   google-cloud-bigquery
    #   great-expectations
    #   openlineage-airflow
    #   openlineage-python
rsa==4.7.2
    # via
    #   awscli
    #   google-auth
ruamel.yaml==0.17.17
    # via great-expectations
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
s3transfer==0.5.0
    # via
    #   awscli
    #   boto3
scipy==1.7.2
    # via great-expectations
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   bleach
    #   google-auth
    #   grpcio
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   openlineage-airflow
    #   openlineage-integration-common
termcolor==1.1.0
    # via great-expectations
terminado==0.12.1
    # via notebook
testpath==0.5.0
    # via nbconvert
tomli==1.2.2
    # via black
toolz==0.11.2
    # via altair
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.62.3
    # via great-expectations
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
typing-extensions==3.10.0.2
    # via
    #   black
    #   jsii
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via great-expectations
urllib3==1.26.7
    # via
    #   botocore
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
widgetsnbextension==3.5.2
    # via ipywidgets
zipp==3.6.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
